+++ src/BOSS/BuildOrderPlotter.cpp
@@ -128,7 +128,7 @@ void BuildOrderPlotter::writeBuildOrderPlot(const std::vector<BuildOrderPlotData
     ss << "boxWidthScale = " << 1.0 << std::endl;
     
     int currentLayer = 0;
-    int currentObject = plots[0].m_buildOrder.size();
+    size_t currentObject = plots[0].m_buildOrder.size();
 
     for (size_t p(0); p < plots.size(); ++p)
     {
